requirements.txt
google-api-python-client # Google API
google-search-results # SerpAPI
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
python-dotenv
langchain
openai
slack-sdk
slack-bolt
flask
gunicorn
urllib3<2
huggingface_hub
pandasai
openai
python-dotenv
requests
matplotlib
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.1
greenlet==2.0.2
aiohttp==3.8.4